requirements.txt
librosa==0.10.2.post1
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.4.1.post1
scipy==1.13.1
torch==2.2.1
torchvision==0.17.1
customtkinter==5.2.2
PyAudio==0.2.14
PySoundFile==0.9.0.post1
sounddevice== 0.4.7
pydub==0.25.1
setuptools==70.0.0
